界面修改 & 双击新增行增强

+++ StudentRatingSystem/StudentInf.h
@@ -4,7 +4,7 @@ typedef struct StudentInf{
 	double	studentID;					// 学号
 	wchar_t	name[5];					// 姓名
 	wchar_t studentClass[20];			// 班级
-	unsigned short mark_subject1;		// 科目1成绩
-	unsigned short mark_subject2;		// 科目2成绩
-	unsigned short mark_subject3;		// 科目3成绩
+	float mark_subject1;		// 科目1成绩
+	float mark_subject2;		// 科目2成绩
+	float mark_subject3;		// 科目3成绩
 } studentInf;
